+++ packages/Ecotone/src/Messaging/Handler/Gateway/GatewayProxyBuilder.php
@@ -32,7 +32,6 @@ use Ecotone\Messaging\Handler\Processor\MethodInvocationProcessor;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\AroundInterceptorBuilder;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\AroundMethodInvoker;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\InterceptorWithPointCut;
-use Ecotone\Messaging\Handler\Processor\MethodInvoker\InterceptableMessageProcessor;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\MethodInterceptorBuilder;
 use Ecotone\Messaging\Handler\Processor\MethodInvoker\PayloadResultMessageConverter;
 use Ecotone\Messaging\MessageHeaders;
@@ -433,17 +432,13 @@ class GatewayProxyBuilder implements InterceptedEndpoint, CompilableBuilder, Pro
                 fn (AroundInterceptorBuilder $aroundInterceptor) => $aroundInterceptor->compileForInterceptedInterface($builder, $interceptedInterfaceReference, $this->endpointAnnotations),
                 $aroundInterceptors
             );
-            $messageProcessorInvocation = new Definition(InterceptableMessageProcessor::class, [
-                $gatewayInternalProcessor,
-            ]);
-            $resultConverter = new Definition(PayloadResultMessageConverter::class, [$interceptedInterface->getReturnType()]);
             $messageProcessorInvocation = new Definition(AroundMethodInvoker::class, [
-                $messageProcessorInvocation,
+                $gatewayInternalProcessor,
                 $compiledAroundInterceptors,
             ]);
             $messageProcessors[] = new Definition(MethodInvocationProcessor::class, [
                 $messageProcessorInvocation,
-                $resultConverter,
+                new Definition(PayloadResultMessageConverter::class, [$interceptedInterface->getReturnType()]),
             ]);
         } else {
             $messageProcessors[] = $gatewayInternalProcessor;
